Revert back to using long long.
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/bfp@117 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/glmBfp/src/types.h
@@ -20,7 +20,7 @@ typedef int Int;
 // general unsigned integer
 typedef unsigned int PosInt;
 // general large unsigned integer
-typedef unsigned long long int PosLargeInt;
+typedef unsigned long int PosLargeInt;
